Temporary fix for color variation id issue
ID format: {parent_product_id}-{color_term_id}

This temporary solution ignores all the ID settings.
Only one hardcoded color attribute is supported.
If is a varient product and does not have the given color variation
attribute the parent product id is used.
If is not a varient product the product id is used.